windows::storage::fileio for file reading;
moved away from fstream-based reading of sgf files; reason being that
opening files that way is subject to standard file permission handling,
which limits you to files stored on external storage; on the other hand,
if you handle them with fileio API, it allows you to open and read files
from the phone internal storage as well. adds a bit of redundant code,
but could easily be hidden in a new method given a necessity (all thanks
to dispatchedhandler being a separate wrapper for a lambda handling ui
changes appropriate for the context).